Move objects/message.{c,h} to new subdirectory
