+++ src/CMakeLists.txt
@@ -70,7 +70,7 @@ set(SOURCE_FILES
     logger/module.c
     main.c
     objects/array.c
-    objects/message.c
+    objects/message/message.c
     objects/object.c
     objects/set.c
     objects/string.c
